Moved the processOne to a blueprint
